refactor(session): move pipeline_factory → session/context/ (PipelineFactory→Factory)
Refs #88
